Modularize comment section, remove extraneous code
